must set a stunnel pidfile location that's writeable by the setuid user
This is unfortunate because the default location is in a rundir:

stunnel[21127]: LOG3[main]: Cannot create pid file /var/run/stunnel4.pid

We have to sacrifice this if we want to run as non-root. Maybe in 18.04,
stunnel will be handled by a genuine systemd unit and we won't need pidfiles?